Fix cloning status on repo page not updating (#19207)
We resubscribed to the of(asError(error)), instead of the outer observable, which only retires if the error is still an error, and doesn't refetch properly. This fixes it by not breaking the observable chain.

I believe this is a regression that was introduced in #12536, but I didn't git bisect to verify.