Convert determineResultTypes to use new result.Types
In many places, we are using `[]string` to represent the set of result
types to be returned. This is a poor representation because it is
difficult to test membership, does not guarantee uniqueness of the
members, and uses way more memory than necessary. Instead, we can use a
bitset, which makes our code more simple and more type-safe.

There are more places where we use string literals, but this was the
place where it caused the most complexity. Once we have a well-defined
base query, we should have the result types be a member on that rather
than being generated from strings.