gitserver: Infer hostname from frontend requests
In order to shard repos across gitserver instance we need a consistent
view of our list of gitserver instances and the hostnames of each
gitserver instance across all frontend and gitserver instances.

Instead of doing this by sharing environment variables we'll keep
frontend as the single source of truth and requests made from it to
gitserver will be used to determine the hostname of each gitserver
instance.

This means that when gitservers first start up they will not know their
own address until they receive their first request from frontend.