search: Add check to avoid OOB panic (#19275)
We were seeing an out of bounds panic in Buildkite on this test. I was
unable to reproduce locally without putting a lot of effort into it, but
it was pretty clear from the stacktrace where this was coming from. This
just adds a simple check to avoid accessing into a nil slice.