search: enable streaming search by default (#19300)
Lots of work went into making this a reality. Add the changelog entry
and flip the default.

We only read the user setting in the webapp, so we only need to update
what we return from the GraphQL resolver. This seems to be the first
time we have changed defaults via what the backend returns, which is why
this PR is bigger than one line.