frontend: Enable uncloned repo count on cloud
This was disabled in the past because it was slow, this change updates
the query used so that we again return the number of uncloned repos for
cloud users.

For normal users it counts any repos added by external services they
own.

For admin users it also inclued repos added by ANY site level external
service, apart from our default GitHub and GitLab repos since they're
cloned on demand.