repo-updater: Stop calling SetClonedRepos
This function sends a list of ALL cloned repos to the database in order
to sync the cloned state from gitserver. We no longer need to do this
since gitserver now maintains the gitserver_repos table.

Listing of default_repos now also joins against this new table instead
of relying on repo.cloned