Don't import from enterprise in OSS web client codebase (#19382)
Seems like we've missed some cases of imports from enterprise in the non-enterprise package.
To prevent this from happening again, I also added a small script to check in CI.

This also has the positive side-effect that things that aren't available in the OSS version won't show in the site-admin UI navigation anymore. They were dead links that linked to a 404 page anyways. (Like the extensions link, codeintel,..)