search: Add prometheus metric for overall search latency
To avoid further increasing the cardinality of the
`src_graphql_field_seconds` metric, this commit creates a new exported
prometheus metric `src_search_response_latency_seconds` which only
tracks the response latency for a single search query.

Additionally, it adds support for parsing the search-blitz query name
from the `User-Agent` string as the source so that we can uniquely
identify named queries to use as sentinels of search performance.