search: increase capacity of bufferedSender (#19440)
We increase capacity of bufferedSender to 4x.

Based on the logs we added in #19358, we can see that
`stream.total_send_time_ms` for `func err patternType:regexp` is around
7 seconds when aggregated across all zoekt instances. The reason is most
likely backpressure created by `bufferedSearcher`. With a capacity of 30
events it can only hold 1-2 events per zoekt instance on Cloud. The
traces show us that we can expect 100+ events for queries such `func err
patternType:regexp`.

Ideally we find a buffer size that is small enough not to cause OOMs,
and large enough so that it fills up just when repos resolution is done.