Use mdi-react icons rather than graphql icons
This commit modifies our result rendering code for commit and repository
results to use a static icon rather than an icon transmitted as part of
the result.

Why? Because currently, for every commit and repository result, we send
a base64-encoded SVG along with it. Based on some simple measurements,
this makes up approximately half of our gzipped payload. Additionally,
it complicates our resolver interfaces to have to deal with icons from
the backend.

This does not remove the icons from the graphQL schema yet, since I
don't know if we can safely do that without breaking backwards
compatibility.