Move sg.config.yaml to repo root & allow private overwrites
This moves the sg.config.yaml to the repository root so that one can
just use `sg start` and `sg run gitserver` etc. without having to
specify the `-conf` flag.

It also adds support for a local `.gitignore`d
`sg.config.overwrites.yaml` file that users can add to the repository.
If it exists, it will then get merged with the basic configuration when
`sg` starts.

What's the usecase? Custom command sets, custom env vars, etc. that you
don't want to commit to the repository.