search: Fix commit search duplicated results issue (#19460)
* search: Fix commit search duplicated results issue

Thic commit fixes an issue where a commit would show up twice in the
results. What was happening is we would first do a search for commits
containing a specific pattern. Then, using those results, we'd make a
second request for the commit content so we can filter on the files
modified by the commit (and maybe get some other things too).

However, the issue was that in the second request, the original query
args were being added to the command. This is problematic because
sometimes, the first query will return incomplete results so we could
get partial results faster. But, since the second query has all the
original query args, it would return _all_ the results, without caring
about the commit OIDs passed in to it.

This commit removes the original query args from the rawShowSearch
function so we only return results for the commit OIDs passed to it.

* Add integration test

* Add changelog entry