database: Improve speed of external service deletion
Instead of scanning the entire repo table looking for orphaned repos
after an external service has been deleted, we instead only scan repos
we know may have been affected.

We had to use a temporary table since if the deletion was done in a CTE
it is not visible in subsequent statements.