frontend: Reduce steps needed to add repositories
When first adding a code host the user is presented with linked to "add
repositories". The link now takes them directly to the manage page where
they can add repos instead of to the "code host connections" page where
they would again have to click on the "add repositories" button.