search: Run each query independently in search-blitz (#19578)
* Run each query independently

This change spins up a goroutine per query so that they each execute on
an independent interval. This way, one long-running query won't block
the other queries, and we can get a more consistent interval between
queries. It also gives us the ability to run queries with an independent
interval.

Note that time.Ticker only "stacks" one tick, so a long-running query
won't cause a flood of queries after it completes.

* search: Configurable intervals for search-blitz (#19579)

* Add configurable intervals to search-blitz

This adds an optional `interval:` configuration to each of our query
configs. This allows us to run queries on independent intervals, so for
expensive or long-running queries, we can make them run more
infrequently.

* Fix lint