internal/repoupdater: Add func NewClient and use for DefaultClient
This commit adds a function NewClient which should take the
responsiblity of initiating a new repoupdater client. The benefit of
using this function is that it allows us to set a custom URL for the
repoupdater service at the time of instantiation.

At the moment we are limited to the env variable REPO_UPDATER_URL
changing which would affect all uses of DefaultClient, especially in
tests.