gitserver: Abort doBackgroundRepoUpdate when rate limiter is zero
If we don't abort then the background job can potentially block forever
if a zero rate limit has been set. This would leak goroutines and hold a
cloneLimiter lock.

We do not add a timeout as it's hard to pick on as this is a legitimate
background process that could take a long time to complete for large
repos.