search: do not dedup within an endpoint
It is possible for two (or more) endpoints to have the same repos
indexed. As such we dedup results such that we only ever return results
for a repo from the same endpoint. However, when we introduced Zoekt
streaming we could call dedup multiple times per endpoint. This lead to
us ignore all results from a repository after the first event. To
reproduce this behaviour you need to do a search on a large repo since
it will have multiple results. For example search "event" in the zig
repo.