search contexts: prevent flickering navbar (#19921)
While we wait for the result of the isSearchContextSpecAvailable call, we assume the context is available
to prevent flashing and moving content in the query bar. This optimizes for the most common use case where
user selects a search context from the dropdown.

If a search context is unavailable (e.g. visiting a link from other users), the flicker will still happen, but it should be largely unnoticeable because the rest of the page will still be loading too.