Use the Select implementation on result types (#19991)
We made result types implement the Select() method, however, we weren't
using it, so that was dead, stale code. This plugs those methods into
the resolver Select() implementations.

It does some duplicate work, but this should be temporary since we
should (hopefully) be eventually selecting on result types and not
resolvers.