oobmigration: Mark all existing migrations as non-deprecated (#19980)
We currently have no way to check whether a required out-of-band migration has completed on startup. There are a few current migrations that mark 3.27 as the deprecation version (meaning that in 3.27 the migrations will no longer run and the migrations must have completed).

This seems incorrect. Two of these are on 0% on Cloud and if this feature were complete, Cloud would be unable to boot.

This just resets the deprecation date and assumes that all the registered migrators were not going to be removed prior to the upcoming branch cut. If this is wrong YELL.