search: introduce basic limiter (#20007)
This adds a limiter to our GraphQL API that limits requests from
anonymous callers. The rate limit is configurable through the
site configuration.

There was already a much more sophisticated, but currently disabled,
rate limiter in place that currently undergoes testing. This new
limiter is a heavily simplified version of the existing limiter which we
can use for the time being.