enterprise: Enforce user and external service limits
We were setting hook functions to check enforced limits in a Init
function which set the hooks on our global database struct and in the
move away from using globals some instances were being used without the
trigger functions.

This is a minimal changes that defines the trigger functions globally
instead.