Translate sentinel query dashboard to our generator (#20066)
This commit translates our sentinel query dashboard into our dashboard
generation tool. That will allow us to hook into the standard monitoring
and alerting infrastructure as well as (maybe eventually) make this a
useful feature for customers.

This wasn't done initially because it was very much in flux as we
decided what metrics and thresholds would be useful. Additionally, I
didn't realize we needed to go through the generator to hook in to our
standard alerting infrastructure.