insights: expose job status info over GraphQL API
This exposes some status/progress information we can use in the Code Insights
frontend to communicate to the viewer what is going on, how many points have
been collected, if errors ocurred, etc.

For sure this isn't perfect, the most beautiful API, always super clear, etc.
but it's a _lot_ better than what we have right now (nothing.)

<img width="783" alt="image" src="https://user-images.githubusercontent.com/3173176/115091535-9bd6e180-9ecc-11eb-9856-ba74784ee8e2.png">

Missing tests, but it's a pretty straightforward implementation so I'm content
merging it as-is (especially given the other option is for us to not have this
until we get a backend engineer.)

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>