Update batch changes changelog for 3.27 release (#20143)
This PR updates the in-app changelog list for Batch Changes.

It also adds the deprecation notice for site-admin changesets being created with external service tokens. This is a very edge-case and a relict from earlier times, so I don't think we need to advertise it more prominently than we do here.

What this deprecation means:

When you created a Batch Change as a site admin before, and never configured user credentials, we fell back to using the code host connection tokens for backwards compatible with the early days of Batch Changes, where only site-admins were able to publish changesets. We ask you to configure a user credential if you haven't done so, or configure a global service account to be used with Batch Changes. All regular users of Batch Changes aren't affected by this deprecation at all.