Upgrade css-loader to latest version
There were a few config changes and a change to the URL resolution algorithm, which was easiest to work around by moving the one only url in the entire webapp outside of the sass code. This should be another precursor for webpack 5 support.
Also I had to pin to-string-loader to a version that fixes usage with css-loader v4+, but was never released to NPM.