Don't require a batch changes resolver in OSS build (#20150)
This PR only includes the Batch Changes schema extension in the GraphQL API, when using an enterprise build. Hence, we don't need the `defaultBatchChangesResolver` anymore, and the API console in OSS isn't cluttered with a lot of inaccessible APIs. This also required refactoring how the node resolver works a bit, so we never call the `BatchChangeByID` resolver, when a user passes a node id like `base64(BatchChange:123)`, but fail with the expected error.