zoekt: indexserver prevent HEAD pointing to master (#20260)
This updates zoekt to include one more commit. Commit description below.

If master is not your default branch, but you are indexing it then we
would not index the default branch. This happened because in the git
repo with synthesize we run `update-ref HEAD foo` then `update-ref
master bar`. Because `HEAD` pointed to `refs/heads/master` we would have
a final value of `bar` for HEAD.

This commit changes the default branch to a random string which
shouldn't be indexed. Now `HEAD` will point to the random string,
preventing this issue.