Increase alert threshold on sentinel queries (#20259)
I doubled the duration the thresholds need to be exceeded
for an alert to be triggered so that now, an alert should only fire if the threshold
is exceeded for more than 30 minutes. This should be plenty of time to
filter out all the minor service blips, and only happen when either there is a
DoS-like event or a regression.