extsvc: store additional metadata in other repo types (#20296)
When building the clone url for other repositories, we store the repo name as the external id, which is not good.
It should be using the path of the clone URL. However, changing the external ID will cause all other repos to be deleted and then added as new repos.

Instead of doing that, we are adding metadata which will have what we want as the path. Then we can take the external service id and concat the metadata. This will also be consistent between src-expose and normal other repos.