Fix regression and e2e test suite (#20338)
For some reason ts-node doesn't pick the tsconfig file that belongs to the source files, but the tsconfig that is in the cwd, which is setting module: esnext, which isn't supported by node. We do want to keep this flag though because it enables better build results with webpack.
I think the ultimate fix would be to move the regression and e2e test suites out of the web codebase into separate ts projects, but that would take a moment and I want to fix CI now.