Fix flaky scheduler.ticker test by checking channel is closed
This should fix #20350.

Since closing a channel (which is what's happening under the hood of
`ticket.stop()`) is *not* a blocking operation, it's possible that the
goroutine of the test closes the channel and then immediately reads from
`ticker.C`. The ticker goroutine then sees the read from `t.C` before
the closing of `done` first.

The fix here leaks a bit of internals by checking `ticker.done`
explicitly. Another way to make this non-flaky would be to always  make
two reads of `ticker.C` and only check the second value.