Replace `SymbolMatch.BaseURI` with new `result.File` (#20310)
This commit replaces the information stored in a `SymbolMatch.BaseURI`
with a reference to its containing file.

The goal of this is to be able to replace the call to
`SymbolMatchResolver.URL()` with a `SymbolMatch.URL()`, but currently, a
`SymbolMatch` does not contain the information needed to create that
URL.

In order to add a reference to its containing file without introducing
circular references between `result.FileMatch` and `result.SymbolMatch`,
the information identifying the file itself was pulled out of
`result.FileMatch` into a new `result.File` and was added to the
`result.SymbolMatch` struct.

Additionally, this commit removes the `gituri` package. The only place
it was being non-trivially used was in `SymbolResult.BaseURI`, and that
no longer exists.