Fix localStorage integration test issue (#20380)
Fixes #20379

Search contexts integration tests fail occasionally. I think this is because we aren't waiting until the page is done loading, so changing Puppeteer to wait until there are no network requests seems to fix the issue. This failure is not super common, but after running 6 builds I didn't see it, so hopefully this actually fixes it.