Remove error from CanonicalURL (#20382)
CanonicalURL() can never fail, but we were propagating an error because we making a URL then parsing it, which can theoretically fail. By passing around *url.URL structs instead, we can ensure there is never a parse failure. Additionally, this will be useful for adding fragments and query params to URLs further down the line.