internal/vcs: Allow @ in repo URL (#20297)
This fix ensures that repos with `@` in their name will now be cloned, but they will still not be searchable from the UI as we need to update our expectations to handle search syntax for revisions. Additionally, opening the repo will also return a 404 since the API splits the repo name on the first `@` at the moment.