Don't drop spans for larger SQL queries
For some sql queries, logging all the arguments could've exceeded the maximum size of 65k bytes of a single span, causing the span not to show up in the trace at all.
I think this is a reasonable tradeoff between being able to introspect the entire query, and being able to get a useful trace.