Increase go test timeout to 5m
There's been a couple of CI runs recently that exceeded the 4min threshold, given that we constantly add new tests, I think this is the right thing to do. (Of course we can and should invest in making the existing tests faster, but this should fix it for the time being).