batches: don't panic on unexpected GitHub PR action
When converting a raw payload to changeset event kind, we need to be
able to handle payloads of unexpected types, especially nil, since we
don't control the inputs that come in via webhooks.

The specific case that triggered this was a GitHub pull request event
that had an unknown action: this results in a nil event keyer being
returned, and that caused a panic in ChangesetEventKindFor().