Fix initial search queries failing after startup of frontend pod
This is just a workaround, because I don't have capacity to speed up the query/write a background cache refiller service.
What this does: It loads the default repos at startup, so the cache is warm when the first search query comes in. The cache is busted after a minute, but it is regularly re-filled by the sentinel queries (I think?), so as a first stopgap for this issue, this should be a temporary fix. Issue XX tracks the progress on this.