Fix stale diff stat of unpublished changesets
Before this change the diff stat of changesets wasn't updated when a new
changeset spec was attached.

While investigating this we also found out that when checking whether
the derived state of a changeset needs to be updated we assume that
"Open" is the only non-final state a changeset can be in. But "draft" is
the other one.

So we also updated that.