Fix out-of-bounds panic for highlighted diffs with trailing delete line
This is the mirror PR to
https://github.com/sourcegraph/sourcegraph/pull/20505 and fixes the same
issue for trailing "-" lines.

The added complication here is that we can't assume that trailing lines
are always the last lines in a hunk. So we need to do an additional loop
to detect where the last delete-line is that's not followed by another
"unchanged" line.

Happy to give this PR away though, but we wanted to fix the panic we ran
into :)