Conditional execution of steps with `if:` (#520)
This extends the schema of the batch spec `steps` to add a new field:

- `if: <string that can use templating to evaluate to "true" or "false">`

(the corresponding schema changes `sourcegraph/sourcegraph` are here: https://github.com/sourcegraph/sourcegraph/pull/20399)

`if:` is evaluated in a best-effort partial-evaluation manner before the execution of the `steps`. If we can statically evaluate it, we use the result of that evaluation to build a per-repository/per-workspace list of tasks, which increases cache utilisation.

See the comments in `partial_eval.go` for more details.

It also extends the `StepContext` in the templates to have the `steps.{added,modified,deleted}_files` and `steps.path` fields.