Fix file start anchor for diff search
This fixes an issue where specifying a file start anchor for a diff
search like `file:^CHANGELOG.md$` would cause the search to return no
matches because we removed the `--no-prefix` flag from the `git show`
command. This adds that flag back, so file names are in the form
`CHANGELOG.md` rather than `a/CHANGELOG.md`.