sg: improve output of install script (#20716)
My heart stopped a little when I ran this on a Linux box and was
cheerily informed that sg had been installed to /usr/bin/sg. In
practice, it was just that I didn't have ~/go/bin in my $PATH.

Let's try to detect this, and then we can even warn the user if `sg`
won't do what they think it will.