Make DefaultRepoStore a cache wrapper (#20645)
It being a store and actually just delegating to repo store, adding a cache layer was confusing, because the general pattern is that you can always create a store from a db handle. Because of that, we had an incident where search latencies went up, because the cache was always busted, after we recreated the store each time it's invoked.