database: User added private repos should be listed (#20727)
Our permissions system was not taking into account private repos added
via code host connections, not sign in connections.

This fix is a workaround that assumes any code added by a user added
external service should be visible by the user who owns the external
service. The assumption is that if they had access to sync the repo,
they should have access to read it. This approach side steps our
existing permissions syncing system and so is not an ideal long term
solution but is quick to implement.