search: handle empty in concat when converting regex to repo names (#20583)
This commit correctly handles a repository name regex like
`^foo(bar|)$`, which represents the two strings `foobar` and
`foo`. Previously we would only match `foo`.

allMatchingStrings takes a regular expression and tries to convert it
into a list of exact strings, or a list of SQL wildcard patterns. This
is a performance optimization that exists to work around postgresql's
lack of optimizations in its regex engine.

The code isn't 100% correct. However, it is more correct than the
previous implementation (famous last words). This function really needs
to be completely rewritten. But we haven't had to tweak it in a very
long time, so I am instead doing a quicker fix of patching this problem
that came up.