Move search code out of graphqlbackend
This takes a first pass at moving search code out of
cmd/frontend/internal and cmd/frontend/graphqlbackend.

The strategy is to move most files as-is into internal/search/run, then
fix compilation errors. Further organization will come in following
commits, but this is the "rip off the band-aid" commit.