Add new "JVM Packages" external services type.
Supersedes #20930.

The goal of this PR is to replace a prototype implementation of RFC 307
called "PackageHub". PackageHub is currently implemented as a `src serve-git`
service type, which has the following limitations:

- Complicated installation: PackageHub is deployed as a separate service.
  Ideally, the functionality of PackageHub should be included with
  Sourcegraph out of the box.
- Slow latency: it takes a long time between when a Java package is
  added to PackageHub and when it becomes available on Sourcegraph.

This PR adds a new external service type called "Maven" that
reimplements one part of PackageHub - the part that proxies
source code that has been published to Maven Central (npm equivalent for
JVM ecosystem). PackageHub additionally supports proxying the sources of
npm package as well as the JDK (Java standard library). Those features
will be added separately in a future PR.

Co-authored-by: Noah S-C <noah@sourcegraph.com>