Don't run TS compiler on dev start (#20902)
The reason this was here was that code intel in the editor works best if you have a fresh compile of the current ts project. However, the VS Code settings we ship include a ts watcher task already so this is redundant and causes additional load on startup, which isn't great. Also, it often causes logspam when people are pulling latest main and someone added a new gql operation and the type gen hasn't finished running before tsc gets to it. I think every engineer who needs this can either use the vscode watcher or set it up themselves and it should not be part of the dev stack starting up.