Fix issue with file filters causing invalid diff results
This commit fixes an issue where adding a `lang:` filter to a diff
search would cause non-matching commits to show up in the results.

This was ultimately caused by the `--max-count` flag on `git show`,
which, very unintuitively, ends up causing `git show` to ignore the
passed-in commit hashes and just list up to the number of commits
specified to `--max-count`. Additionally, I could not find any
references to using `--max-count` with `git show`, so my guess is it's
just not a supported argument for `git show`.

Really, though, we just don't need to be appending the format args to
the `git show` command at all, since we are just getting more
information for a specific set of commits. So, this commit removes those
args entirely.