Implement actions on changeset lists (#20361)
This PR implements the bulk operations tab and header notifications, and the UI element to trigger the create comment operation. Therefor, I made the checkboxes available on all list views and migrated the detach operation to just be another case of a bulk operation. I will convert it into a proper bulk operation in the backend in a follow-up PR.