Protect from highlighting panic
New highlighting changes uncovered a lot of complexity in how we handle
trailing newlines. We've made quick attempts to correct this before, but
were still hitting more edge cases. In attempt to at least stop the
panics from occuring, this commit wraps all accesses to highlighted code
that might have had newlines truncated in a `safeIndex` function, which
will return a default value in the cases where it would otherwise have
panicked with an out-of-bounds error.