validate settings JSON for sideloaded extensions (#20783)
Previously, the settings JSON Schema of a sideloaded extension was not passed to the Monaco settings editor. This meant that settings did not validate or complete using the sideloaded extension's schema. For example, if you had an extension that added a new property `foo`, you wouldn't see that property or its docs in the Monaco completion menu.

Now, the sideloaded extension's JSON Schema is passed to the Monaco editor (along with that of all the enabled extensions from the registry).