search: use streaming search protocol when creating patterns (#21196)
The commit which introduced ToTextPatternInfo (a2f7bc9772e) never passed
in the streaming protocol. This meant that when a user did not specify a
limit, we would use the defaults for batch search.

This had some bad interactions, since all the layers above the specific
backends would only think a limit was hit when we had 500 or more
results. But the backends would stop searching after 30. This lead to
very strange result counts that would change on every search.

This has been a regression since v3.27 series. We have a known work
around (specify a limit), but I believe its a relatively bad user
experience. As such this may justify a patch release on the 3.27
series (and not just be destined for 3.28).