apidocs: add barebones web UI for internal testing  (#21224)
This PR adds a barebones web UI. It will be behind a feature flag and enabled
for only the Sourcegraph team for now, as it is not up to our quality standard.

While there are numerous aspects that could be improved here, I think the code
quality is OK and I haven't done anything egregious here - I've followed our
existing code / patterns as closely as possible.

One notably missing component here is tests. Since it is more important to reach
an MVP of API docs than have it be well-tested, we are intentionally deferring
that until later. I am tracking every case that we need tests for, including this
PR, here: #20567

Depends on #21223

Fixes #19397
Helps #19389

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>