frontend: warm default repo cache before serving traffic (#21238)
Pre-heat default repo cache. This loads the default repos at startup, so
the cache is warm when the first search query comes in. This query can
take ~20s on sourcegraph.com, so this avoids users hitting a cold
cache. Issue #20651 tracks removing this again.

We also do some cleanup in frontend's main function.

Co-authored-by: Erik Seliger <erikseliger@me.com>