Add fuzzy file finder with keyboard shortcut `t`. (#20921)
Previously, there was no way to interactively search for a file that
matches a given fuzzy query. This commit adds a new modal that activates
on the keyboard shortcut `t` and allows the user to quickly open a file
that fuzzily matches a given query.

The implementation in this commit is specifically designed to handle
large repositories with >100k source files. The fuzzy finder achieves
low-latency by:

1) implementating all the filtering inside the browser on the
   client-side
2) using bloom filters to quickly discard buckets of filenames that are
   guaranteed to not match a given query.  This technique is used by the
   Scala language server and documented in this blog post
   https://scalameta.org/metals/blog/2019/01/22/bloom-filters.html

The downside of this approach:

1) it can take a while to first download all filenames in the
   repository, especially for large repos like chromium/chromium. It should be
   possible to optimize some parts of this on the backend.
2) the bloom filter algorithm is sometimes more strict compared to the
   fuzzy finder in VS Code and IntelliJ. There are techniques to make
   the fuzzy finder more fuzzy, for example we could make all-lowercase
   queries case-insensitive. It's also worth consider whether we want
   use a different filtering algorithm for smaller repositories (<25k source files)