add limitations page to code insights docs (#21384)
The purpose of this page is to pre-empt disappointment and support requests. It's essentially the docs.sourcegraph version of the "considerations" section of our original gist.

When communicating with customers we haven't spoken to directly before about code insights (so they haven't heard about the limitations from me or a member of the team), this page will help set expectations in a clear and consistent way.