Use an OS entropy source for random seed (#21416)
We were getting conflicts in database name, which seems to be because we
were getting identical nanoseconds back from `time.Now()`. This the RNG
seed to use an OS source of entropy (from crypto/rand) as the source so
we can be more sure that different test packages don't happen to hit
the same nanosecond.