servegit: decompress input to git-upload-pack (#547)
Ensure that if the Content-Encoding of a servegit request is gzip then
we read it as a compressed input. git-upload-pack does not accept
gzipped input and fails. This is only sometimes gzipped and I'm not sure
of the conditions so we can't rely on it being always or never encoded.

This is a port of sourcegraph/sourcegraph 8e1e58c.

Co-authored-by: Eric Fritz <eric@sourcegraph.com>