Recycle test databases
Creating a new database with every call to NewDB() was putting a lot of
contention on postgres system tables. In slower environments like CI,
this was causing the tests to time out even though it allowed for much
faster tests locally. This commit adds test database recycling, in which
tests try to pass their database on to another test before destroying
it.