searcher: run comby tests in parallel (#21388)
We noticed that searcher tests took a few seconds locally. This was
surprising, so we investigated and it seems it is dominated by the cost
of starting comby multiple times. This PR moves to using parallel
testing to hide some of the costs. Part of this was moving away from
autogold in one test to allow parallel subtests.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>