search: remove truncating of zoekt line and fragments
We had some old logic which limited the number of line matches and line
fragments we would return from zoekt. It was based on the magic variable
"k", which was a scaling number which is used to guess how many
results. So this logic existed to return less results per file if we
expected a lot of results.

This breaks user expectations as well as breaking exhaustive search. It
breaks exhaustive search since we would never return all matches if a
file contained too many matches on a line or too many lines with
matches.

It also leads to subtle bugs where small changes to your query can lead
to totally different results returned due to this truncating interacting
with display limits.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>