restore `experimentalFeatures.copyQueryButton` as deprecated (#21463)
The feature flag was originally removed in #21364. However, this is causing errors in some cases ([see Slack thread](https://sourcegraph.slack.com/archives/CHEKCRWKV/p1622133020136000)). Since the description of `experimentalFeatures` explicitly says `Features that are now enabled by default are marked as deprecated`, I am restoring the flag and adding a deprecation message.

Eventually, we should be able to remove deprecated feature flags if #21462 is fixed.